refactor: isolate heat pump state from Modbus
