+++ main/heatpump_params_screen.cpp
@@ -8,8 +8,8 @@
 
 #include "heatpump_control_screen.h"
 #include "nav_bar.h"
-#include "modbus/arctic_heatpump.h"
-#include "modbus/arctic_registers.h"
+#include "heatpump_controller.h"
+#include "heatpump_types.h"
 #include "macon_state.h"  // arctic::setpoint_limits / SetpointKind
 #include "macon_advanced_params.h"  // arctic::AdvancedParam, AP reg map
 #include "advanced_params.h"        // advanced_param_read/write (controller IO)
