+++ main/heatpump_screen.cpp
@@ -10,8 +10,8 @@
 #include "heatpump_errors.h"
 #include "event_log_screen.h"
 #include "nav_bar.h"
-#include "modbus/arctic_heatpump.h"
-#include "modbus/arctic_registers.h"
+#include "heatpump_controller.h"
+#include "heatpump_types.h"
 #include "app_preferences.h"
 #include "i18n/i18n.h"
 #include "ui_common.h"
